+++ CHANGELOG.md
@@ -414,3 +414,4 @@ This update overhauls Predator's dashcam functionality and adds various features
     - Predator can automatically apply an offset to the current time if the system time drifts from the GPS time.
 - Improved configuration validation.
 - Added `state.json` interface file to communicate Predator's current mode of operation to external programs.
+- Added offline push notifications to dash-cam mode via Reticulum LXMF.